# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Food Ordering Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, confirmed, preparing, ready, delivered, cancelled]
        - name: dateFrom
          in: query
          required: false
          schema:
            type: string
        - name: dateTo
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/payment:
    patch:
      operationId: updateOrderPaymentStatus
      tags: [orders]
      summary: Update order payment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderPaymentStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/rider:
    patch:
      operationId: assignRider
      tags: [orders]
      summary: Assign or unassign a rider to an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiderAssignment"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /riders:
    get:
      operationId: listRiders
      tags: [orders]
      summary: List delivery riders
      responses:
        "200":
          description: List of riders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rider"
    post:
      operationId: createRider
      tags: [orders]
      summary: Create a new rider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rider"
  /riders/{id}:
    patch:
      operationId: updateRider
      tags: [orders]
      summary: Update a rider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiderInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rider"
    delete:
      operationId: deleteRider
      tags: [orders]
      summary: Delete a rider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
...
components:
  schemas:
...
    OrderItem:
      type: object
      required: [menuItemId, name, quantity, unitPrice]
      properties:
        menuItemId:
          type: integer
        name:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
    Order:
      type: object
      required: [id, customerName, customerPhone, status, total, items, createdAt, paymentMethod]
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, delivered, cancelled]
        total:
          type: number
        notes:
          type: ["string", "null"]
        paymentMethod:
          type: string
          enum: [cash_on_delivery, jazz_cash, easy_paisa]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
    OrderInput:
      type: object
      required: [customerName, customerPhone, items, paymentMethod]
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: string
        notes:
          type: string
        paymentMethod:
          type: string
          enum: [cash_on_delivery, jazz_cash, easy_paisa]
        items:
          type: array
          items:
            type: object
            required: [menuItemId, quantity]
            properties:
              menuItemId:
                type: integer
              quantity:
                type: integer
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, delivered, cancelled]
...
    Rider:
      type: object
      required: [id, name, phone, active]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        active:
          type: boolean
        restaurantId:
          type: ["integer", "null"]
    RiderInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        active:
          type: boolean
    OrderPaymentStatusUpdate:
      type: object
      required: [paymentStatus]
      properties:
        paymentStatus:
          type: string
          enum: [pending, confirmed, not_required]
    RiderAssignment:
      type: object
      properties:
        riderId:
          type: ["integer", "null"]
...