# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Food Ordering Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderPaymentMethod } from './orderPaymentMethod';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  customerName: string;
  customerPhone: string;
  /** @nullable */
  customerAddress?: string | null;
  status: OrderStatus;
  total: number;
  /** @nullable */
  notes?: string | null;
  paymentMethod: OrderPaymentMethod;
  items: OrderItem[];
  createdAt: string;
}