# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Food Ordering Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputItemsItem } from './orderInputItemsItem';
import type { OrderInputPaymentMethod } from './orderInputPaymentMethod';

export interface OrderInput {
  customerName: string;
  customerPhone: string;
  customerAddress?: string;
  notes?: string;
  paymentMethod: OrderInputPaymentMethod;
  items: OrderInputItemsItem[];
}