# lib/api-zod/src/generated/types/orderPaymentStatusUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Food Ordering Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderPaymentStatusUpdatePaymentStatus } from './orderPaymentStatusUpdatePaymentStatus';

export interface OrderPaymentStatusUpdate {
  paymentStatus: OrderPaymentStatusUpdatePaymentStatus;
}